codegen::recursion: physical detector move out of lean/
Move the proof-mode recursion classifier from `codegen::lean` into a
new shared module `codegen::recursion::detect`. The previous wrapper
in `codegen::recursion::analyze_plans` delegated back to the Lean
module; now the detector body lives in the shared module and Lean
calls through the same public entry that Dafny already uses.

Moved verbatim into `recursion/detect.rs` (backend-neutral AST
pattern matching only): `collect_calls_from_{expr,body}`,
`expr_to_dotted_name`, `call_matches`/`call_is_in_set`/
`canonical_callee_name`/`call_matches_any`, `is_ident`,
`is_int_{minus,plus}_positive`, `is_skip_ws_{advance,same}`,
`is_string_pos_advance`, `StringPosEdge` + `classify_string_pos_edge`,
`ranks_from_same_edges`, `is_scalar_like_type`,
`collect_list_tail_binders{,_from_expr}`,
`recursive_constructor_binders`,
`grow_recursive_subterm_binders_from_expr`,
`collect_recursive_subterm_binders`,
`single_int_countdown_param_index`, `has_negative_guarded_ascent`,
`single_int_ascending_param`, `extract_equality_bound_expr`,
`single_list_structural_param_index`,
`supports_single_sizeof_structural`,
`supports_single_string_pos_advance`,
`supports_mutual_int_countdown`,
`supports_mutual_string_pos_advance`,
`supports_mutual_sizeof_ranked`, and the top-level entry
`proof_mode_recursion_analysis` (renamed to `analyze_plans`).

Stays in `codegen::lean` (bumped to `pub(crate)` so the shared
detector can call in): `pure_fns`, `recursive_type_names`,
`recursive_pure_fn_names`, `find_type_def`,
`sizeof_measure_param_indices`. These still lean on
`toplevel::is_pure_fn` / `is_recursive_type_def` / `type_def_name`,
which are Lean-private helpers over the AST. Moving them to a
shared AST-utility module is a sensible second pass but deliberately
out of scope here. `recurrence::detect_second_order_int_linear_recurrence`
was already `pub(crate)` and is called from the shared detector via
`crate::codegen::lean::recurrence::...`; the `recurrence` submodule
visibility was bumped from private to `pub(crate)` accordingly.

`bound_expr_to_lean` stays in `codegen::lean` as Lean-specific
syntax — only Lean consumes it.

Pure refactor: no behaviour change. All 470 lib tests pass. Proof
sweep unchanged: quicksort Lean 22 verify 2 sorry, quicksort Dafny
22/8, fibonacci Dafny 22/2, law_auto 0 sorry, oracle_trace 0 sorry,
json Dafny 96/132.
